+++ src/App.js
@@ -8,7 +8,7 @@ import Agregar from './components/Pages/Agregar';
 import { Router, Route, Redirect } from 'react-router';
 import history from './history';
 
-import Ensamble from './components/Forms/Ensamble';
+import Ensamble from './components/Ensamble/Ensamble';
 
 import './App.css';
 
@@ -23,7 +23,7 @@ const App = () => {
     <Fragment>
       <Router history={history}>
         <Navbar />
-        <div className="main">
+        <div className='main'>
           <Route path='/' exact component={() => <Redirect to='/inicio' />} />
           <Route path='/inicio' exact component={Home} />
           <Route path='/cotizar' exact component={Quote} />
