+++ src/components/Ensamble/EnsambleContext.js
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+export const EnsambleContext = createContext(null);
